feat(web-server): Scan mode validator + small refactor

+++ backend/web-server/controllers/scan-mode.controller.ts
@@ -1,45 +1,52 @@
 import { KoaContext } from '../koa';
 import { ScanModeCommandDTO, ScanModeDTO } from '../../../shared/model/scan-mode.model';
+import ValidatorInterface from '../../validators/validator.interface';
 
-const getScanModes = async (ctx: KoaContext<void, Array<ScanModeDTO>>) => {
-  const scanModes = ctx.app.repositoryService.scanModeRepository.getScanModes();
-  ctx.ok(scanModes);
-};
+export default class ScanModeController {
+  constructor(private readonly validator: ValidatorInterface) {}
 
-const getScanMode = async (ctx: KoaContext<void, ScanModeDTO>) => {
-  const scanMode = ctx.app.repositoryService.scanModeRepository.getScanMode(ctx.params.id);
-  ctx.ok(scanMode);
-};
-
-const createScanMode = async (ctx: KoaContext<ScanModeCommandDTO, void>) => {
-  const command: ScanModeCommandDTO | undefined = ctx.request.body;
-  if (command) {
-    const scanMode = ctx.app.repositoryService.scanModeRepository.createScanMode(command);
-    ctx.created(scanMode);
-  } else {
-    ctx.badRequest();
+  async getScanModes(ctx: KoaContext<void, Array<ScanModeDTO>>): Promise<void> {
+    const scanModes = ctx.app.repositoryService.scanModeRepository.getScanModes();
+    ctx.ok(scanModes);
   }
-};
 
-const updateScanMode = async (ctx: KoaContext<ScanModeCommandDTO, void>) => {
-  const command: ScanModeCommandDTO | undefined = ctx.request.body;
-  if (command) {
-    ctx.app.repositoryService.scanModeRepository.updateScanMode(ctx.params.id, command);
-    ctx.noContent();
-  } else {
-    ctx.badRequest();
+  async getScanMode(ctx: KoaContext<void, ScanModeDTO>): Promise<void> {
+    const scanMode = ctx.app.repositoryService.scanModeRepository.getScanMode(ctx.params.id);
+    if (scanMode) {
+      ctx.ok(scanMode);
+    } else {
+      ctx.notFound();
+    }
   }
-};
 
-const deleteScanMode = async (ctx: KoaContext<void, void>) => {
-  ctx.app.repositoryService.scanModeRepository.deleteScanMode(ctx.params.id);
-  ctx.noContent();
-};
+  async createScanMode(ctx: KoaContext<ScanModeCommandDTO, void>): Promise<void> {
+    try {
+      await this.validator.validate(ctx.request.body);
 
-export default {
-  getScanModes,
-  getScanMode,
-  createScanMode,
-  updateScanMode,
-  deleteScanMode
-};
+      const scanMode = ctx.app.repositoryService.scanModeRepository.createScanMode(ctx.request.body as ScanModeCommandDTO);
+      ctx.created(scanMode);
+    } catch (error: any) {
+      ctx.badRequest(error.message);
+    }
+  }
+
+  async updateScanMode(ctx: KoaContext<ScanModeCommandDTO, void>): Promise<void> {
+    try {
+      await this.validator.validate(ctx.request.body);
+      ctx.app.repositoryService.scanModeRepository.updateScanMode(ctx.params.id, ctx.request.body as ScanModeCommandDTO);
+      ctx.noContent();
+    } catch (error: any) {
+      ctx.badRequest(error.message);
+    }
+  }
+
+  async deleteScanMode(ctx: KoaContext<void, void>): Promise<void> {
+    const scanMode = ctx.app.repositoryService.scanModeRepository.getScanMode(ctx.params.id);
+    if (scanMode) {
+      ctx.app.repositoryService.scanModeRepository.deleteScanMode(ctx.params.id);
+      ctx.noContent();
+    } else {
+      ctx.notFound();
+    }
+  }
+}
